--- HW/achornyi/TASolution/UI/Pages/CatalogPage.py
@@ -1,5 +1,0 @@
-from HW.achornyi.TASolution.UI.Pages.BasePage import BasePage
-
-
-class CatalogPage(BasePage):
-    pass
