--- HW/achornyi/TASolution/UI/Pages/SystemPage.py
@@ -1,5 +1,0 @@
-from HW.achornyi.TASolution.UI.Pages.BasePage import BasePage
-
-
-class SystemPage(BasePage):
-    pass
