+++ HW/achornyi/ta_solution/ui/components/base_page/header.py
@@ -1,0 +1,3 @@
+class Header:
+    def __init__(self):
+        pass
