+++ HW/achornyi/ta_solution/ui/pages/base_page.py
