+++ HW/achornyi/ta_solution/ui/pages/system_page.py
@@ -1,0 +1,5 @@
+from HW.achornyi.ta_solution.ui.pages.base_page import BasePage
+
+
+class SystemPage(BasePage):
+    pass
